fix potential problem with always increasing cache size

+++ mobile/src/main/java/net/fred/feedex/utils/NetworkUtils.java
@@ -36,14 +36,13 @@ import net.fred.feedex.provider.FeedData;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
-import java.io.FilenameFilter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.CookieHandler;
 import java.net.CookieManager;
 import java.net.HttpURLConnection;
 import java.net.URL;
-import java.util.regex.Pattern;
+import java.util.HashSet;
 
 import okhttp3.OkHttpClient;
 import okhttp3.OkUrlFactory;
@@ -116,23 +115,36 @@ public class NetworkUtils {
         }
     }
 
-    public static synchronized void deleteEntriesImagesCache(Uri entriesUri, String selection, String[] selectionArgs) {
+    public static synchronized void deleteEntriesImagesCache(long keepDateBorderTime) {
         if (IMAGE_FOLDER_FILE.exists()) {
-            PictureFilenameFilter filenameFilter = new PictureFilenameFilter();
 
-            Cursor cursor = MainApplication.getContext().getContentResolver().query(entriesUri, FeedData.EntryColumns.PROJECTION_ID, selection, selectionArgs, null);
+            // We need to exclude favorite entries images to this cleanup
+            Cursor cursor = MainApplication.getContext().getContentResolver().query(FeedData.EntryColumns.FAVORITES_CONTENT_URI, FeedData.EntryColumns.PROJECTION_ID, null, null, null);
+            if (cursor != null) {
+                HashSet<Long> favIds = new HashSet<>();
+                while (cursor.moveToNext()) {
+                    favIds.add(cursor.getLong(0));
+                }
 
-            while (cursor.moveToNext()) {
-                filenameFilter.setEntryId(cursor.getString(0));
-
-                File[] files = IMAGE_FOLDER_FILE.listFiles(filenameFilter);
+                File[] files = IMAGE_FOLDER_FILE.listFiles();
                 if (files != null) {
                     for (File file : files) {
-                        file.delete();
+                        if (file.lastModified() < keepDateBorderTime) {
+                            boolean isAFavoriteEntryImage = false;
+                            for (Long favId : favIds) {
+                                if (file.getName().startsWith(favId + ID_SEPARATOR)) {
+                                    isAFavoriteEntryImage = true;
+                                    break;
+                                }
+                            }
+                            if (!isAFavoriteEntryImage) {
+                                file.delete();
+                            }
+                        }
                     }
                 }
+                cursor.close();
             }
-            cursor.close();
         }
     }
 
@@ -236,27 +248,5 @@ public class NetworkUtils {
         connection.connect();
 
         return connection;
-    }
-
-    private static class PictureFilenameFilter implements FilenameFilter {
-        private static final String REGEX = "__[^\\.]*\\.[A-Za-z]*";
-
-        private Pattern mPattern;
-
-        public PictureFilenameFilter(String entryId) {
-            setEntryId(entryId);
-        }
-
-        public PictureFilenameFilter() {
-        }
-
-        public void setEntryId(String entryId) {
-            mPattern = Pattern.compile(entryId + REGEX);
-        }
-
-        @Override
-        public boolean accept(File dir, String filename) {
-            return mPattern.matcher(filename).find();
-        }
     }
 }
